bugfix: incorrect latencies due to overlapping
The previous check for if we received an "ack" for our "ping"
only checked srcip and srcport. If we receive multiple "acks",
perhaps due to a retransmit, then the latency appears to be much
lower than it actually is.

We solve this problem by ensuring that the response synack's acknum
is seq+1, where seq is the seq num we sent in the packet. This ensures
that we are only measuring the reply for the packet we actually sent,
and not a previous one.